// src/features/customerList/__generated__/FILTER_BOAT_TYPE_LABELS.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: FILTER_BOAT_TYPE_LABELS
// ====================================================

export interface FILTER_BOAT_TYPE_LABELS_boatTypes {
  __typename: "BoatTypeType";
  id: string;
  name: string | null;
}

export interface FILTER_BOAT_TYPE_LABELS {
  boatTypes: FILTER_BOAT_TYPE_LABELS_boatTypes[] | null;
}