// src/features/customerList/__generated__/FILTER_WINTER_STORAGE_PLACE_LABEL.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: FILTER_WINTER_STORAGE_PLACE_LABEL
// ====================================================

export interface FILTER_WINTER_STORAGE_PLACE_LABEL_winterStoragePlace {
  __typename: "WinterStoragePlaceNode";
  id: string;
  number: number;
}

export interface FILTER_WINTER_STORAGE_PLACE_LABEL {
  winterStoragePlace: FILTER_WINTER_STORAGE_PLACE_LABEL_winterStoragePlace | null;
}

export interface FILTER_WINTER_STORAGE_PLACE_LABELVariables {
  winterStoragePlaceId: string;
}